Refactor the "eval*" methods of TreeNodeType to obtain **kwargs For some options that should be applied in a whole evaluation.
In most cases, creating a new operator is better because usually you
want two functions with minor differences to be invoked in an evaluation
instead of making all functions listen to the same config.

So when is this extension useful?
The previous commit 72e013ffd3f29be268282673f9a199493b34e36f shows an
attempt to make a weird variable have a default value, but only in the
variation of AssignOperator.
However, if we want all the weird variables in an expression to have
default values, we did need to create variations that handle anonymous
variables to "EVERY" operator we use.
It is much simpler to just have an option in evaluation that tells
calcs.types.Var to create temporary vars into the mapping, so we can
easily use the feature for all operators in one specific evaluation.

If anyone has a demand that something should be done in some specific
evaluations onto all of specific operators, then that's it.